Modified index.html
I was seeing some errors in the chrome browser related to the height of the canvas, so removed the default `height=500` from the inline declaration.
Please do not forgot to check in the Mozilla browser too, cuz the project is built in the perspective of the mozilla browser.

Update (via git commit --amend): After checking, the height should not be removed instead it should be set less than 500, let us say to 300.